HSEARCH-746 Move implementations of SearchConfiguration as private
