+++ hibernate-search/src/main/java/org/hibernate/search/event/FullTextIndexEventListener.java
@@ -62,7 +62,7 @@ import org.hibernate.search.Version;
 import org.hibernate.search.backend.Work;
 import org.hibernate.search.backend.WorkType;
 import org.hibernate.search.backend.impl.EventSourceTransactionContext;
-import org.hibernate.search.cfg.SearchConfigurationFromHibernateCore;
+import org.hibernate.search.cfg.impl.SearchConfigurationFromHibernateCore;
 import org.hibernate.search.engine.AbstractDocumentBuilder;
 import org.hibernate.search.engine.SearchFactoryImplementor;
 import org.hibernate.search.spi.SearchFactoryBuilder;
